requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
crypto==1.4.1
cryptography==37.0.2
idna==3.3
Naked==0.1.31
pycparser==2.21
python-decouple==3.6
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
urllib3==1.26.9
